requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
boto3==1.28.48
botocore==1.31.48
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==5.0.0
Django==4.2.5
django-admin-rangefilter==0.11.1
django-cors-headers==4.2.0
django-filter==24.1
django-import-export==3.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.6
importlib-metadata==7.1.0
importlib-resources==5.13.0
inflect==7.0.0
jmespath==1.0.1
Markdown==3.6
MarkupPy==1.14
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
paramiko==3.3.1
pillow==10.3.0
pycodestyle==2.11.0
pycountry==23.12.11
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
tqdm==4.66.2
typing-extensions==4.7.1
urllib3==1.26.16
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
zipp==3.18.1
